spruce up sign in sign up pages
