+++ apps/nextjs/src/app/(main)/players/[username]/page.tsx
@@ -1,3 +1,4 @@
+import { Suspense } from "react";
 import { getProfile } from "~/app/server-actions/profiles/actions";
 import ComingSoon from "~/components/coming-soon";
 
@@ -16,14 +17,23 @@ export async function generateMetadata(
   return { title: player?.username ?? "Player" };
 }
 
-export default async function PlayerProfilePage(
+export default function PlayerProfilePage(
   props: Readonly<PlayerProfilePageProps>,
 ) {
+  return (
+    <Suspense fallback={null}>
+      <PlayerProfilePageSuspense {...props} />
+    </Suspense>
+
+  )
+}
+
+async function PlayerProfilePageSuspense(props: PlayerProfilePageProps) {
   const params = await props.params;
-  const player = await getProfile(params.username);
+  // const player = await getProfile(params.username);
 
   return (
-    <ComingSoon title={player?.username ?? "Player Profile"}>
+    <ComingSoon title={ `${params.username}'s Profile` }>
       <h2>The Player Profiles are under construction</h2>
     </ComingSoon>
   );
