--- apps/nextjs/src/app/(main)/sign-in/[[...sign-in]]/page.tsx
@@ -1,17 +1,0 @@
-"use client";
-
-import { SignIn, useUser } from "@clerk/nextjs";
-
-export default function SignInPage() {
-  const { user } = useUser();
-
-  if (user) {
-    return <div>You are signed in.. </div>;
-  }
-
-  return (
-    <div className="flex h-full min-h-screen w-full items-center justify-center">
-      <SignIn />
-    </div>
-  );
-}
