--- apps/nextjs/src/app/(main)/sign-up/[[...sign-up]]/page.tsx
@@ -1,13 +1,0 @@
-"use client";
-
-import { SignUp, useUser } from "@clerk/nextjs";
-
-export default function SignUpPage() {
-  const { user } = useUser();
-
-  if (user) {
-    return <div>You are signed in.. </div>;
-  }
-
-  return <SignUp path="/sign-up" routing="path" />;
-}
